refactor(native-lib): modularize Rust bindings into ffi/result/streaming modules
Extract monolithic lib.rs (586 lines) into focused modules following
feat/native-lib-language-wrappers architecture while preserving all
critical safety fixes from fix/rust-go-bindings:

- src/ffi.rs: GraalVM types, extern "C" declarations, isolate lifecycle
- src/result.rs: ExecutionResult type with decode helpers
- src/streaming.rs: Callback contexts, stream iterator, transform logic
- src/error.rs: Migrated to thiserror derive macros with expanded variants
- src/lib.rs: Public API surface (~170 lines), re-exports, encode_inputs

Critical safety invariant preserved:
  unsafe impl<T: Sync> Send for SendPtr<T>

All 12 integration tests pass unchanged.
