+++ native-lib/rust/Cargo.toml
@@ -9,6 +9,8 @@ base64 = "0.22"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 libc = "0.2"
+thiserror = "1.0"
+once_cell = "1.19"
 
 [dev-dependencies]
